requirements.txt
opencv-python==4.11.0.86
pandas == 2.0.3
numpy == 2.3.1
pillow == 11.2.1
pytesseract == 0.3.13
google-cloud-vision == 3.10.2
pdf2image == 1.17.0
PyPDF2 == 3.0.1